requirements.txt
# Django
# ------------------------------------------------------------------------------
django
django-environ
djangorestframework
django-querycount==0.8.3
drf-spectacular==0.26.4
django-filter
psycopg2
# Flake
# ------------------------------------------------------------------------------
flake8
flake8-absolute-import
flake8-comprehensions
flake8-django
flake8-print
flake8-quotes
# GIT
# ------------------------------------------------------------------------------
pre-commit